fix missing configuration and description
