+++ .env
@@ -1,4 +1,4 @@
 MASTER_PW=74RuqICjDPQEFREmhIFaqRf6H
 POSTGRES_USER=postgres
-DOCKER_DNS_NAME=localhost
+localhost=localhost
