+++ docker-compose.override.yml
@@ -13,14 +13,14 @@ services:
       API_POSTGRESQL_HOST: localhost      
       API_POSTGRESQL_USER_NAME: ${POSTGRES_USER}
       API_POSTGRESQL_USER_PASSWORD: 74RuqICjDPQEFREmhIFaqRf6H
-      API_OIDC_URL: http://${DOCKER_DNS_NAME}:8080/realms/Calendar
+      API_OIDC_URL: http://localhost:8080/realms/Calendar
       API_USE_SWAGGER: true
       API_SWAGGER_CLIENT_SECRET: hiTq3F2ALbOiray4ZH4Kq0ON3GVTkHjR
-      API_SWAGGER_REDIRECT_URL: http://${DOCKER_DNS_NAME}:4000/swagger/index.html
-      API_SWAGGER_OIDC_URL: http://${DOCKER_DNS_NAME}:8080/realms/Calendar
-      API_JWT_AUDIENCE: http://${DOCKER_DNS_NAME}:8080/realms/Calendar
-      API_JWT_AUTHORITY: http://${DOCKER_DNS_NAME}:8080/realms/Calendar
-      API_JWT_METADATA_ADDRESS: http://${DOCKER_DNS_NAME}:8080/realms/Calenda
+      API_SWAGGER_REDIRECT_URL: http://localhost:4000/swagger/index.html
+      API_SWAGGER_OIDC_URL: http://localhost:8080/realms/Calendar
+      API_JWT_AUDIENCE: http://localhost:8080/realms/Calendar
+      API_JWT_AUTHORITY: http://localhost:8080/realms/Calendar
+      API_JWT_METADATA_ADDRESS: http://localhost:8080/realms/Calenda
       API_DEBUG_TEST_ENDPOINT_ENABELD: true
       API_DEBUG_TEST_ENDPOINT_POLICY: EDITOR_VIEWER
       TZ: Europe/Berlin
@@ -46,9 +46,9 @@ services:
     command:
       - "--verbose"
       - "start-dev"
-      # - "--hostname-admin-url=http://${DOCKER_DNS_NAME}:8080"
-      # - "--hostname=${DOCKER_DNS_NAME}"
-      - "--hostname-url=http://${DOCKER_DNS_NAME}:8080"
+      # - "--hostname-admin-url=http://${localhost}:8080"
+      # - "--hostname=${localhost}"
+      # - "--hostname-url=http://localhost:8080"
       - "--import-realm"
       - "--features=token-exchange"
     environment:
