+++ docker-compose.yml
@@ -33,10 +33,10 @@ services:
       context: ./src/calendar-client
       dockerfile: Dockerfile
       args:
-        - VITE_API_URL=http://${DOCKER_DNS_NAME}:5000
+        - VITE_API_URL=http://localhost:4000
         - VITE_OIDC_CLIENT_SECRET=Wo9T9nS0ebJbUpVso6wpOGgVluQaqajA
-        - VITE_OIDC_AUTHORITY=http://${DOCKER_DNS_NAME}:8080/realms/Calendar
-        - VITE_OIDC_REDIRECT_URL=http://${DOCKER_DNS_NAME}:3000
+        - VITE_OIDC_AUTHORITY=http://localhost:8080/realms/Calendar
+        - VITE_OIDC_REDIRECT_URL=http://localhost:3000
     restart: always
     depends_on:
       keycloak:
