+++ src/calendar-backend/Calendar.Api/Configurations/JwtEnvironmentConfiguration.cs
@@ -7,10 +7,6 @@ public class JwtEnvironmentConfiguration : IEnvironmentConfiguration<JwtEnvironm
     public string? JWT_AUDIENCE { get; set; } = "http://localhost:8080/realms/calendar";
     public string? JWT_AUTHORITY { get; set; } = "http://localhost:8080/realms/calendar";
     public string JWT_METADATA_ADDRESS { get; set; } = "http://localhost:8080/realms/calendar/.well-known/openid-configuration";
-    [Obsolete]
-    public string? JWT_ISSUER { get; set; }
-    [Obsolete]
-    public string JWT_KEY { get; set; }
     /// <inheritdoc/>
     public JwtEnvironmentConfiguration Validate()
     {
