+++ src/calendar-backend/Calendar.Api/appsettings.json
@@ -5,31 +5,5 @@
       "Microsoft.AspNetCore": "Warning"
     }
   },
-  "Keycloak": {
-    "realm": "Calendar",
-    "verify-token-audience": true,
-    "ssl-required": "none",
-    "resource": "account",
-    "rolesSource": "Realm",
-    "credentials": {
-      "secret": "vTtq8mz0OKiUGLaQWpkNZdpdjLhYPsZ5"
-    },
-    "confidential-port": 0
-  },
   "AllowedHosts": "*",
-  "serilog": {
-    "MinimumLevel": "Verbose",
-    "WriteTo": [
-      {
-        "Name": "Console",
-        "outputTemplate": "[{Timestamp:HH:mm:ss} {Level:u3}] {Message:lj} Properties={Properties:j}{NewLine}{Exception}"
-      },
-      {
-        "Name": "Seq",
-        "Args": {
-          "serverUrl": "http://seq:5341"
-        }
-      }
-    ]
-  }
 }
